# Performance improvements in scaleminmax